feat(proxy): http now working, but still lacking
There are some things that still need to be done:

- Support for requests bigger than the buffer size (for file uploads and
      whatnots)
- Support for responses bigger than the buffer size
- Support for Keep-Alive http connections

With those out of the way, this will **then** be done for and ready for
refactoring and then begin to make the main part of the program

I still need to find a way to get https source codes. Do I need to act
as a SSL/TLS server for the client and a client for the server to get
unencrypted data? (because that'll be a ***pain in the ass***)